A lot of work.
Ground out a lot of the work to get a functional robot. Need to refine
it to more precisely navigate the ball area. Improve hardware so that
ball can be more easily detected as well as push the ball over more
easily. Also, use the light sensor for better localization.