Bandith could work on this code...
...but he probably won't read it. Some things have been smoothed out,
but the bump sensor does need to be put on the mux, the light sensor
taken from the mux and put on IN_3, and in general things could be
smoothed out. It would be good to maintain obstacle avoidance during all
actions, back up into the white after capturing a ball, making a 1/8
turn to get into the black better, and maybe other things.